[chore] domain 폴더 분리
