+++ backend/src/main/java/com/poscodx/pofect/domain/essentialstandard/EssentialStandard.java
@@ -1,5 +1,6 @@
-package com.poscodx.pofect.domain.entity;
+package com.poscodx.pofect.domain.essentialstandard;
 
+import com.poscodx.pofect.domain.entity.BaseEntity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
