+++ backend/src/main/java/com/poscodx/pofect/domain/user/User.java
@@ -1,6 +1,7 @@
-package com.poscodx.pofect.domain.entity;
+package com.poscodx.pofect.domain.user;
 
 
+import com.poscodx.pofect.domain.entity.BaseEntity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
